% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchExACCall.R
\name{getVariantInfoFromExACAPI}
\alias{getVariantInfoFromExACAPI}
\title{Batch Call to the Broad Institue's ExAC project API to get Variant Annotations}
\usage{
getVariantInfoFromExACAPI(codeVector)
}
\arguments{
\item{codeVector}{Character vector of codes consists of the variant
information in a format ("CHROMOSOME-POSITION-REFERENCE-VARIANT")
required for Broad Institue's ExAC project API.}
}
\value{
The output is a list object containing the
variant annotation information from the Broad Institute's ExAC
Project for each variant supplied.
}
\description{
This function make a batch call to the Broad Institue's ExAC project API
to retreive the annotation information of variants.
}
\examples{
library(VCFAnnotatorTempus)
postVariantJson <- getVariantInfoFromExACAPI(c("14-21853913-T-C","14-21854139-A-G","22-46615746-A-G"))
print(postVariantJson[["14-21853913-T-C"]]$variant$allele_freq)
}
\author{
Ashish Jain
}
