% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchExACCall.R
\name{getVariantInfoFromExACAPI}
\alias{getVariantInfoFromExACAPI}
\title{Annotate the variants supplied in a VCF file}
\usage{
getVariantInfoFromExACAPI(codeVector)
}
\arguments{
\item{codeVector}{Character vector of codes consists of the variant
information in a format ("CHROMOSOME-POSITION-REFERENCE-VARIANT")
required for Broad Institue's ExAC project API.}
}
\value{
The output is a list object containing the
variant annotation information from the Broad Institute's ExAC
Project.
}
\description{
Annotate the variants supplied in a VCF file
}
\author{
Ashish Jain
}
